Refactor tests for SubtasksService and response creation
- Moved  tests for filtering, pagination, and grouping for the `CreateSubtaskResponseForTodoItem` method to a new file `CreateSubtasksResponseTests.cs`.
- Refactored namespaces for better organization under `KinaUnaProgenyApi.Tests.Services.Subtasks`.
- Cleaned up unused `using` directives and ensured necessary ones are included in both test files.
